Finish cleanup of commands.ts
