+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "forcecode",
-  "version": "3.17.0-3",
+  "version": "3.17.0-4",
   "publisher": "JohnAaronNelson",
   "displayName": "ForceCode",
   "description": "Visual Studio Code extension for Salesforce (SFDC) development",
@@ -567,7 +567,7 @@
     "@types/universal-analytics": "0.4.2",
     "@types/xml2js": "0.4.4",
     "compressing": "1.4.0",
-    "deepmerge": "3.2.0",
+    "deepmerge": "3.2.1",
     "fs-extra": "8.0.1",
     "globule": "1.2.1",
     "jsforce": "1.8.5",
